Fix update routes (#17)
* fix update inventory route.

* add space

* change type.

* add update detail.

* fix delete that inventory existance check place.

* fix update inventory

* fix create order

* seperate interface of order and order detail.

* fix delete order

* fix update

+++ src/repositories/inventoryRepository.ts
@@ -9,6 +9,8 @@ import {
   selectInventoryItem,
   createInvntoryTable
 } from "../db/invnetoryQueries";
+import { INVALID_INVENTORY_ID_ERROR } from "../constants";
+import { resolvePlugin } from "@babel/core";
 
 class InventoryTable {
   static createInventory(name: String, description: String, price: String, quantityAvailable: String): Promise<void> {
@@ -39,26 +41,38 @@ class InventoryTable {
     );
   }
 
-  static updateInventory(
+  static async updateInventory(
     id: number,
-    name: string | null = null,
-    description: string | null = null,
-    price: number | null = null,
-    quantityAvailable: number | null = null
+    name: string | null,
+    description: string | null,
+    price: number | null,
+    quantityAvailable: number | null
   ): Promise<void> {
     const db = dBSqlite3();
-    return new Promise((resolve, reject) =>
+    if (!this.isInventoryExist(id)) throw INVALID_INVENTORY_ID_ERROR.type;
+
+    new Promise((resolve, reject) =>
       db.run(updateInventoryItem(id, name, description, price, quantityAvailable), (_: RunResult, err: Error | null) =>
         err ? reject(err) : resolve()
       )
     );
   }
 
-  static deleteInventry(id: Number): Promise<void> {
+  static async deleteInventry(id: number): Promise<void> {
     const db = dBSqlite3();
-    return new Promise((resolve, reject) =>
+    if (!this.isInventoryExist(id)) throw INVALID_INVENTORY_ID_ERROR.type;
+
+    new Promise((resolve, reject) =>
       db.run(deleteInventoryItem(id), (_: RunResult, err: Error | null) => (err ? reject(err) : resolve()))
     );
+  }
+
+  private static async isInventoryExist(id: number): Promise<boolean> {
+    const db = dBSqlite3();
+    const inventory = await new Promise<any>((resolve, reject) =>
+      db.get(selectInventoryItem(id), (err: Error | null, _inventory: any) => (err ? reject(err) : resolve(_inventory)))
+    );
+    return inventory !== undefined;
   }
 }
